SERA basic mode
In its basic mode, SERA is an SR-aware firewall that can apply the normal firewall
processing to the original packets even if they have an SR based SFC encapsulation.

Each received packet goes through an sr6_pre_processor() that splits traffic into SR
and non-SR traffic. Non-SR traffic is processed as in an SR-unaware firewall.
SR traffic follows a different path through the firewall. In this  path, the firewall
evaluates source and destination information of the defined rules on the original packet,
properly taking into account the impact of the SR encapsulation.

It supports both encap and insert mode, which implies that the original IPv6 source and
destination information of received packets may be encoded differently as follows:
    - Encap mode: original source and destination are the ones of the inner packet.
    - Insert mode: packets have only one IPv6 header. The original source information is
      in the source address of the IPv6 header, while the original destination is encoded as
      the last SID in the SRH.

The inner_match() function is responsible for getting the original source and destination
information from SR packets and compare them to the defined rules. Once a packet hits a
condition of a rule, the associated standard action (ACCEPT, DROP, etc.) is triggered on
that packet.

The "ip6t_seg6" sysctl is used to enable/disable the basic mode of SERA.