Add support for native mode of SERA
SERA stands for SEgment Routing Aware firewall. It's an advanced
SR aware firewall, with extended matching and actions capabilities.

We introduce the native mode of SERA. In this mode, SERA works as
a legacy firewall (iptables and ip6tables), but with ability to apply
the same set of rules defined for the legacy firewall directly to the
inner packet of SR traffic.

It can handle SR traffic encapsulated in encap mode with inner IPv6
as well as inner IPv4 packets.

IPv6 and IPv4 packets are SR-encapsualted as shown below. To make
a concrete example, if an existing rule includes a condition on the
source and destination address, the condition should be checked on
the source address of the inner packet.

Original IPv6 packet
+=============+====================+
| IPv6 header |       payload      |
+=============+====================+

SRv6 packet
+-------------+---------+=============+====================+
| IPv6 header |   SRH   | IPv6 header |       payload      |
+-------------+---------+=============+====================+

Original IPv4 packet
+==================================+
| IPv4 header |       payload      |
+==================================+

SRv6 packet
+-------------+---------+=============+====================+
| IPv6 header |   SRH   | IPv4 header |       payload      |
+-------------+---------+=============+====================+

An important use case scenario is to virtualize the legacy firewalls,
executing them in servers on the NFV infrastructure, without changing
the legacy rules and with no need of SR-proxy functionality.

Signed-off-by: Ahmed Abdelsalam <amsalam20@gmail.com>