Add support for IPv6 segment routing 'SEG6' target
Service Function Chaining (SFC) is one of the main use-cases of IPv6
Segment Routing (SRv6)[1].

The Segment Routing Header (SRH) allows including a list of segments
in the IPv6 packet. This segment list can be used to steer the packet
though a set of Virtual Network Functions (VNFS) e.g., Firewall, DPI, IDS

The edge node (SR ingress node) classifies the traffic and consequently
includes the segment lists in the IPv6 packet.

The IPv6 segment routing 'SEG6' target supports a set of SR-specific
actions. This set of actions allows branching within a pre-defined
SRv6 SFC policy instead of the usual linear exploration of the VNF chain.

As an example, consider a service chain which includes a firewall
followed by an Intrusion Detection System. The firewall may decide
to skip the IDS for a subset of traffic that matches some conditions.

This patch implements the following SR-specific actions:
(1) go-next: is the default action of the SEG6 target. It sends packets
towards the next SID from SRH.
(2) skip-next: skips the next SID in the SRH.
(3) go-last: skips the remaining part of the segment list and send the
packet towards the last segment.

[1] https://tools.ietf.org/html/draft-clad-spring-segment-routing-service-chaining-00

Signed-off-by: Ahmed Abdelsalam <amsalam20@gmail.com>